requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   django
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
channels==4.0.0
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
cryptography==42.0.1
    # via pyjwt
defusedxml==0.8.0rc2
    # via python3-openid
dj-rest-auth==4.0.1
    # via -r requirements.in
django==5.0.1
    # via
    #   channels
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==0.54.0
    # via -r requirements.in
django-cors-headers==4.3.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.6
    # via
    #   requests
    #   yarl
msgpack==1.0.7
    # via channels-redis
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3.post1
    # via djangorestframework
redis==5.0.1
    # via channels-redis
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sqlparse==0.4.4
    # via django
urllib3==2.1.0
    # via requests
yarl==1.9.2
    # via aiohttp
